refactor: move learn to components
