+++ src/components/learn/learnObjects/wordsArray.js
